+++ src/lib/index.ts
@@ -1,8 +1,8 @@
 export { JSONSchema, JSONSchemaType } from './src/service/validation/json-schema';
-export { AsyncLocalDatabase } from './src/service/databases/async-local-database';
+export { LocalDatabase } from './src/service/databases/local-database';
 export { IndexedDBDatabase } from './src/service/databases/indexeddb-database';
 export { LocalStorageDatabase } from './src/service/databases/localstorage-database';
 export { MockLocalDatabase } from './src/service/databases/mock-local-database';
 export { JSONValidator } from './src/service/validation/json-validator';
-export { ALSGetItemOptions, AsyncLocalStorage } from './src/service/lib.service';
-export { AsyncLocalStorageModule } from './src/module';
+export { LSGetItemOptions, LocalStorage } from './src/service/lib.service';
+export { LocalStorageModule } from './src/module';